feat: Introduce `machine_value` in MachineEvent Model (WEB-3702)
This commit extends the `MachineEvent` model by adding a new attribute called `machine_value` of type JSON. The `machine_value` attribute is used to store the value of the machine state associated with the event.

In the `State` class, the `setEventBehavior` method is updated to include the `machine_value` attribute in the creation of a new `MachineEvent`. The `machine_value` is set to an array containing the current state definition ID. This allows for tracking the state value associated with each event in the history.

Additionally, the `EventStoreTest.php` test case is modified to include an `id` property in the machine definition configuration. This is done to ensure that the `machine_id` attribute in the `MachineEvent` table is properly populated.

The `MachineEventFactory` and `create_machine_table.php.stub` files are also updated to reflect the changes in the `MachineEvent` model.