refactor(EventBehavior): Simplify source type validation (WEB-3703)
This commit simplifies the source type validation in the `EventBehavior` class. The changes include the addition of the `#[WithoutValidation]` attribute to the `$source` property.

In the original implementation, the `$source` property was defined with the `SourceType` enum type and assigned a default value of `SourceType::EXTERNAL`. However, with the introduction of the `#[WithoutValidation]` attribute, the source type validation is skipped for the `$source` property. This attribute indicates that the property should not be validated when creating or updating instances of the `EventBehavior` class.