refactor(MachineDefinition): Rename property `activeStateDefinition` to `currentStateDefinition` (WEB-3748)
This commit refactors the `MachineDefinition` class by renaming the property `activeStateDefinition` to `currentStateDefinition`. The `currentStateDefinition` represents the current state of the machine actor.

The previous property name, `activeStateDefinition`, was not accurately describing its purpose and could be misleading. By renaming it to `currentStateDefinition`, the property's intent becomes clearer and aligns better with its role in representing the current state of the machine actor.