fix(State): Fix context serialization in State class (WEB-3760)
This commit fixes an issue with the serialization of the `context` property in the `State` class.

Previously, the `context` property was being serialized using the `data` property of the `ContextManager` class. However, this caused an error as the `ContextManager` class does not have a `data` property.

The fix involves serializing the `context` property using the `toArray()` method of the `ContextManager` class, which correctly converts the context data into an array.