refactor(State): Initialize `$history` property with a new database collection (WEB-3760)
This commit refactors the `State` class to initialize the `$history` property with a new collection using the `newCollection()` method of the `MachineEvent` class.

Previously, the `$history` property was initialized with an empty collection using the `collect()` helper function. However, by utilizing the `newCollection()` method, we ensure that the `$history` property is always initialized as a fresh instance of a database collection.