test(PersistingStateTest): Add more assertions on state restoration (WEB-3760)
New lines of code were added to the PersistingStateTest. Two new uses are added for EventDefinition and StateDefinition. The changes in the tests are necessary because the test was not covering some important aspects of the state of the restored machine, specifically its context and currentEventBehavior. Now, tests ensure that the machine is restored to the exact same state it was, covering all the necessary parameters."