test: Add history validation to PersistingStateTest (WEB-3760)
Added the validation of MachineEvent history in the PersistingStateTest class and also made a revision in MachineActor.php to return a new state with the updated machineEvents as a history instead of an empty collection. This modification ensures we are correctly associating historical event data with the state in our tests, which provides a more accurate representation of the system status.