feat: Implement Castable interface in EventMachine class (WEB-3768)
Two changes were done in EventMachine class. Firstly, Castable interface is implemented to allow EventMachine instances to be casted to/from primitive types during Eloquent model interaction. Secondly, a new method 'castUsing' was added, which provides the name of the caster class to use during the casting. The MachineCast class is assigned for this purpose. This improves the interoperability of the EventMachine with Eloquent models.