test: Update PersistingStateTest to include non-persistence option (WEB-3768)
This commit modifies the 'INC' events in tests for persisting and restoring machine state to include an option 'shouldPersist' and set it as false. It’s done to ensure that these increments do not persist state changes. Also, a new test has been added to check auto persistence of changes in case 'shouldPersist' is true. This adds flexibility in testing different scenarios and ensures the system behaves correctly both when changes should and should not persist.