fix(HasMachines): Fix attribute existence check
This commit fixes an issue in the `HasMachines` trait where the attribute existence check was not correctly performed. The check now verifies if the attribute is not set and if the cast class is a subclass of `EventMachine`. This ensures that the attribute is only processed if it is not already set and if the cast is an `EventMachine`.

Co-authored-by: Fatih AYDIN <aydinfatih52@gmail.com>