chore(ModelA): Update machine cast definitions (WEB-3770)
This commit updates the machine cast definitions in the `ModelA` class. Previously, the cast definitions for the `abc_mre` and `traffic_mre` attributes were set to their respective machine classes directly. With this update, the cast definitions now include the context key separated by a colon (`:`).

The `abc_mre` attribute is cast using the `AbcMachine` class, and the context key is set to `'modelA'`. Similarly, the `traffic_mre` attribute is cast using the `TrafficLightsMachine` class, and the context key is also set to `'modelA'`.

These changes provide more context and flexibility when using machine actors in the `ModelA` class.

Co-authored-by: Fatih AYDIN <aydinfatih52@gmail.com>