refactor: Update ContextManager set function (WEB-3781)
Improved the set function in ContextManager to now include storing a value either in the 'data' array or as a property depending on the class of the object. Previously, it only stored the value in the 'data' array. This change ensures more fluid handling of the values for various class instances.

Co-authored-by: Fatih AYDIN <aydinfatih52@gmail.com>