feat(TransitionDefinition): Add error message handling for GuardBehavior (WEB-3782)
This commit introduces enhancements to the handling of GuardBehavior in TransitionDefinition.php. Previously, when a guard failed, there was no provision to display an error message.

Now, if a guard fails, it's checked whether it is an instance of GuardBehavior. If true, the associated errorMessage is retrieved and added to the payload of the internal GUARD_FAIL event. This would allow for easier debugging of guard failures, by providing more context about what went wrong and where.