feat(MachineActor): Add option to throw exceptions on guard failure (WEB-3782)
Added an extra flag "$shouldThrowOnGuardFail" to the "send" method in "MachineActor.php". This allows for throwing a "MachineValidationException" when a guard fails. The functionality was added to better handle failure situations in state transitions and provide more control and feedback during execution. Tests updated in "GuardedActionsTest.php" to accommodate this change.