refactor: Update timestamp format in MachineEvent
Changed the timestamp format in the MachineEvent invocation within the Actor's State class. Previously, we were directly using now() method which returned a timestamp in a different format. But certain processes or services consuming this data expected datetime in a specific format. This small tweak of using toDateTime() ensures we are consistent with that expectation and it eliminates any potential errors or complications in data processing downstream.