feat: Add check to MachineCast get method for initializing machines
In the MachineCast class, the get method has been updated to include a check to ascertain whether a machine should be initialized from the model. This is identified using the `HasMachines` trait. This check will prevent inadvertent instantiation of irrelevant or unnecessary machines.

Additional Lines:

A method named `shouldInitializeMachine` has been added to the `HasMachines` trait to facilitate this check. This method returns `true` by default but can be overridden in individual models as necessary.