feat(ContextManager): Add type checking to `has()` method (WEB-3803)
This commit enhances the `has()` method in the `ContextManager` class.
Previously, this method only checked if a key exists in the context.
Now, it can also check if the value associated with the key is of a
specified type. If a type is not provided, the method only checks for
the existence of the key.

Test cases in `ContextManagerTest` have also been added to verify
this new functionality.