feat: Update ContextManager set method to return value (WEB-3691)
The 'set' method in ContextManager class has been updated to return the value that was set. Previously, it had a void return type, but had code to return null if the data was of 'Optional' instance. This change rectifies the issue by changing void to mixed return type. The associated tests have also been updated to reflect the return of the set value.