feat: Allow passing of multiple action arguments in MachineDefinition (WEB-3815)
This commit updates the MachineDefinition class to support passing multiple arguments to an action behavior. Previously, only the state's context and an event behavior were passed to the action behavior. Now, a set of action arguments can be provided and will be passed along when the action behavior is invoked.