test: Add argument array to action methods in tests (WEB-3815)
Added an optional argument array to the __invoke methods of the ActionBehaviors within the traffic lights test stubs. This change was made to provide flexibility in passing additional parameters if necessary in future implementation. It follows the principle of forward compatibility as we are planning to introduce argument passing in action methods.