feat(TransitionBranch): Add description field (WEB-3816)
In the TransitionBranch class, a new description field has been added. This field is set based on the 'description' values in the transitionBranchConfig array. This change was made to provide a mechanism for having more descriptive data associated with transition branches. This newly added description field is also tested in GuardedActionsTest to ensure it works as expected.