Add functionality for actions to return values (WEB-3823)
Two test cases have been added in the ActionTest file to check if actions can return values. A new method 'multipleWithItselfAction' was introduced with this functionality. The ActionBehavior interface has also been updated to reflect this change in the functioning of actions. This additional functionality was needed to provide more flexibility in the use of actions within the EventMachine framework.