feat: Add state matching method to State class (WEB-3824)
A new public method 'matches' was added to the State class in Actor/State.php file. This method checks if the passed string value matches the current state of the machine. If the passed value does not start with the machineId, it prepends it before comparing. This change will facilitate checking if a certain state is active in the state machine. In addition, a corresponding test was added to the StateTest.php test suite.