feat: Add event queuing capabilities to InvokableBehavior's (WEB-3825)
Added functionality to InvokableBehavior that allows actions to raise events. This is accomplished via an event queue in MachineDefinition, which stores raised events during the execution of a transition. A test is added to confirm that actions can raise events appropriately. Action classes XAction, YAction, and ZAction were created as a part of the interactions tested. This new feature is designed to increase flexibility in event-complex environments and enhance the functionality of the InvokableBehavior class.