feat: Add JSON serialization to MachineActor and test (WEB-3841)
The MachineActor class now implements JsonSerializable, enabling it to represent an object in serialized format for use in ModelA attribute serialization. Additionally, a new test file, MachineActorSerializationTest.php, is added to ensure the correct functionality of this new serialization ability. The ModelA class has also been updated to use HasFactory for testing convenience. This improves our testing capabilities and the flexibility in working with MachineActor instances.