feat: Add string casting to MachineActor class (WEB-3846)
Implemented string casting feature for the MachineActor class. The newly added method __toString() will now return a string representation of the current object. This change is for situations where we might need a string representation of the object, for example, during debugging or logging. A corresponding test for this feature was added in MachineActorCastingTest.