feat: Add context requirement check to GuardBehavior (WEB-3851)
Added a new private method hasMissingContext to check if the machine context has any required context attributes defined in GuardBehavior. This function is invoked before any guard behavior execution to ensure the required machine context for the behavior is provided, if missing it throws a MissingMachineContextException. Additionally, new tests have been added to verify this functionality. This helps in enforcing the requirement checks for context attribute mandatory for the execution of the guard behavior.