feat!: Refactor `hasMissingContext` method to static (WEB-3691)
The method `hasMissingContext` in TransitionDefinition.php was refactored to be a static method. This was done to enable direct method access without having to instantiate an unnecessary object of TransitionDefinition. This change enhances efficiency and potentially improves performance. The method has also been moved to a separate static methods region for better code organization and readability. No changes have been made to the functionality of the method.