feat!: Change access modifier of hasMissingContext to public (WEB-3691)
Earlier, the method hasMissingContext was protected and only accessible within its class and its subclasses. It has been modified to be a public method, which is now accessible from anywhere. This change is needed to allow external classes to check if there are any missing context attributes, enhancing modularity and flexibility of our module.