refactor!: Remove unused code in GuardBehavior.php (WEB-3691)
The unused method, validateRequiredContext, and its related imports, TransitionDefinition and MissingMachineContextException, have been removed from the GuardBehavior.php in the src/Behavior directory. This method was previously used to validate the required context for a given guard behavior and context manager, however it's no longer needed. This contributes to code cleanliness and maintainability.