feat: Add exception handling for non-existing transitions (WEB-3864)
The changes introduce error handling for event transitions that do not exist in the machine state definition. A new test case has been added to ensure that `NoTransitionDefinitionFoundException` is thrown when an unknown event is invoked. Changes are noted in files `TransitionDefinitionTest.php` and `MachineDefinition.php`. This elimination of silent failures assists in debugging and transparency in machine state transitions.