feat: Add NoStateDefinitionFoundException for undefined transitions (WEB-3864)
This commit adds a new Exception type, NoStateDefinitionFoundException. This exception will be thrown when a state transition definition is lacking for a certain event type within the current state definition. This is added to ensure that every event type has a defined state transition, maintaining the integrity of the state machine's logic.